final/requirements.txt
colorama==0.4.6
PyQt5==5.15.9
PyQt5_sip==12.13.0
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
python_nmap==0.7.1
Requests==2.31.0
urllib3==2.2.1
